Fixed argument passing to the ldap keyword.
Without this change, the argument connection was simply ignored
because the syntax for the "plugin_args" is "plugin_args(@_)", not
simply "plugin_args".

Which, btw, is totally useless as D2 plugins wants the Dancer2
namespace.

This change comes with two test files which use a public LDAP server
for testing.

The first file 01-bare-ldap.t will try to contact the public box and
write out a ldap.conf file in the t/ directory, which in turn will be used
by our tests in t/ldap.t